abstract shared functions
