+++ apps/extension/src/lib/background/handlers/__tests__/vaultHandlers.test-utils.ts
@@ -1,0 +1,7 @@
+import type { vi } from 'vitest'
+
+export function captureKeeperMessage(): Record<string, unknown> | undefined {
+  const calls = (chrome.runtime.sendMessage as ReturnType<typeof vi.fn>).mock
+    .calls
+  return calls[0]?.[0] as Record<string, unknown> | undefined
+}
