+++ apps/extension/src/lib/background/handlers/__tests__/vaultHandlers.test.ts
@@ -10,6 +10,7 @@ import {
   handleLock,
 } from '@/lib/background/handlers/vaultHandlers'
 import type { VaultMessage } from '@/lib/background/types'
+import { captureKeeperMessage } from './vaultHandlers.test-utils'
 
 vi.mock('@/lib/background/handlers/authHandlers', () => ({
   checkPendingAuthAfterUnlock: vi.fn(),
@@ -41,12 +42,6 @@ function stubKeeperBridge(keeperResponse: unknown = { ok: true }) {
       lastError: undefined,
     },
   } as unknown as typeof chrome
-}
-
-function captureKeeperMessage(): Record<string, unknown> | undefined {
-  const calls = (chrome.runtime.sendMessage as ReturnType<typeof vi.fn>).mock
-    .calls
-  return calls[0]?.[0] as Record<string, unknown> | undefined
 }
 
 describe('handleLock', () => {
