+++ packages/shared/src/auth/stores/__tests__/authStoreTestMocks.ts
@@ -1,4 +1,5 @@
 import { SUI_TESTNET_CHAIN } from '@mysten/wallet-standard'
+import { User } from 'oidc-client-ts'
 import type { Mock } from 'vitest'
 import { vi } from 'vitest'
 import { makeJwt } from '#/testing'
@@ -293,4 +294,34 @@ export function makeTokenResponse(): OAuthTokenResponse {
     expires_in: 3600,
     expires_at: 4600,
   } as OAuthTokenResponse
+}
+
+export const FUTURE = Math.floor(Date.now() / 1000) + 3600
+
+export function makeStoredJwt(overrides: Record<string, unknown> = {}) {
+  return {
+    id_token: makeJwt({ sub: 'user-1', iat: 1000, exp: FUTURE }),
+    access_token: 'at',
+    token_type: 'Bearer',
+    scope: 'openid',
+    refresh_token: 'rt',
+    expires_in: 3600,
+    expires_at: FUTURE,
+    ...overrides,
+  }
+}
+
+export function makeUser(
+  overrides: Partial<ConstructorParameters<typeof User>[0]> = {},
+) {
+  return new User({
+    id_token: makeJwt({ sub: 'user-1', iat: 1000, exp: FUTURE }),
+    access_token: 'at',
+    token_type: 'Bearer',
+    scope: 'openid',
+    refresh_token: 'rt',
+    profile: { sub: 'user-1' } as User['profile'],
+    expires_at: FUTURE,
+    ...overrides,
+  })
 }
